requirements.txt
anyio==4.9.0
backtrader==1.9.78.123
beautifulsoup4==4.13.4
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
exceptiongroup==1.2.2
frozendict==2.4.6
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
multitasking==0.0.11
numpy==2.2.4
pandas==2.2.3
peewee==3.17.9
platformdirs==4.3.7
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.0
pytz==2025.2
requests==2.32.3
schedule==1.2.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
yfinance==0.2.55
